refactored and added support for attribute filters

+++ commercetools.Sdk/commercetools.Sdk/Product/Attributes/NumberAttribute.cs
@@ -4,7 +4,7 @@ using System.Text;
 
 namespace commercetools.Sdk.Domain
 {
-    public class NumberAttribute : Attribute<double>
+    public class NumberAttribute : Attribute<int>
     {
         // TODO See which type should be used exactly
     }
